requirements.txt
absl-py==1.0.0
appnope==0.1.2
asgiref==3.4.1
asttokens==2.0.5
astunparse==1.6.3
backcall==0.2.0
beautifulsoup4==4.10.0
black==21.12b0
bootstrap4==0.1.0
boto3==1.20.33
botocore==1.23.33
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
debugpy==1.5.1
decorator==5.1.1
dj-database-url==0.5.0
Django==4.0.1
django-bootstrap4==21.2
django-heroku==0.3.1
entrypoints==0.3
executing==0.8.2
flatbuffers==2.0
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
gunicorn==20.1.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.10.0
ipython==8.0.0
jedi==0.18.1
jmespath==0.10.0
joblib==1.1.0
keras==2.7.0
Keras-Preprocessing==1.1.2
libclang==12.0.0
Markdown==3.3.6
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
nest-asyncio==1.5.4
numpy==1.22.0
oauthlib==3.1.1
opt-einsum==3.3.0
pandas==1.3.5
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.1
prompt-toolkit==3.0.24
protobuf==3.19.3
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.11.2
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
pyzmq==22.3.0
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.8
s3transfer==0.5.0
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==1.4.29
sqlparse==0.4.2
stack-data==0.1.3
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-cpu==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
threadpoolctl==3.0.0
tomli==1.2.3
tornado==6.1
traitlets==5.1.1
typing_extensions==4.0.1
urllib3==1.26.8
wcwidth==0.2.5
Werkzeug==2.0.2
whitenoise==5.3.0
wrapt==1.13.3
zipp==3.7.0
